create register.js
